Added C shared library and python library.

+++ signalprocessing.h
@@ -1,3 +1,3 @@
 // type: 0=minmax; 1=statistical; 2=absolute.
-extern void imgscale(double* img, int size, int param1, int param2, int type);
+extern void imgscale(double* img, int size, double param1, double param2, int type);
